Update psee-streamer for the new IP interface specification
This aligns with the IPs in the 1.0.0 Kria release.
Unfortunately, the register layout changes too much to even identify
we're dealing with an old IP, thus this breaks support of previous
versions of the IPs.
This is deemed acceptable since those are FPGA IPs only present on
Kria, and fixable through software update, along with this driver.

Signed-off-by: Christian Dreher <cdreher@prophesee.ai>