tkeep-handler: declare AXI4S clocks in the required DT properties
It may be optional, but the driver requires it, and it will always be
present. Unlike Xilinx IPs, there is a common clock for AXI-lite and
AXI-stream interfaces, in case this changes someday, I arbitrarily
define this as the AXI-stream. That way the description stays the same
as the generic one (pass-through).

Signed-off-by: Christian Dreher <cdreher@prophesee.ai>