packetizer: declare AXI4S clocks in the required DT properties
The driver requires it, thus this does not break anything, and fixes
the driver (by changing the spec it must comply with). The clock is
common for streaming interface and register accesses, but, in case
this ever changes, the clock is defined as the stream one, as it makes
more sense to use this one to compute time slices of event stream.

Signed-off-by: Christian Dreher <cdreher@prophesee.ai>