Rework pipeline_start_stop to stop sources first
The previous implementation expected the sub-devices to clear their
internal memories between stop and start, but some did not, resulting
in garbage data at the beginning of streams, coming from the end of
the previous stream. This change hides it by stopping data producers
first, leaving some time for the receivers to finish transfers
gracefully.

As the pipeline keeps a reference to its output, stopping it last
requires to keep a back-track from the sensor to the DMA, this is done
using a recursive function and the call stack. Opportunistically, this
rework expand the design to support sink pads that are not index 0,
and possibly several active sources for a given sub-device, allowing to
merge information in some nodes.

Signed-off-by: Christian Dreher <cdreher@prophesee.ai>